perf(bezier_root,step_time): migrate to f32 for hardware FPU on Cortex-M4F
The F446 (180 MHz Cortex-M4F, BTT Octopus bottom MCU) has a hardware FPU
for single-precision only — `__aeabi_dmul` software-emulates f64 at
~50-100 cycles per operation. With PRODUCER_BATCH_CAP=32 Newton solves
per producer fire × ~7,800 cycles per emulated solve, the producer's
callback exceeded ~1.4 ms on F4, triggering Klipper's
`armcm_timer.c:152` shutdown check ("Rescheduled timer in the past")
when the dispatcher re-read `timer_read_time` after the callback
returned.

The f32 migration:
- `bezier_root.rs`: de Casteljau, Newton, and bisection all in f32.
  Both M4F and M7 have hardware f32 FPU (single-cycle multiply-add),
  cutting per-solve cost from ~7,800 cycles to ~80 cycles on F4 (~100×).
- `step_time.rs::StepTimeQuery`: CPs, t_curr, t_segment_end, and
  step_distance all f32. `compute_next_step_time` returns f32 `t`.
- `engine.rs::cubic_piece`: returns f32 CPs/knots (matches storage —
  `cv.control_points: &[f32]`, was previously upcast to f64).
- `engine.rs::producer_step` & `arm_step_timer`: f64 ↔ f32 cast at
  the producer-state boundary (`ps.t_resume()` stays f64 for
  cross-call continuity; we down-cast on read, up-cast on write).
  Cycle math (`dt_cycles = u_global * duration_f64`) stays f64 so
  multi-second segments at MHz clocks stay exact.

Precision impact:
- f32 ulp at 300 mm bed scale = 3.6e-5 mm. New EPS_CONVERGENCE =
  1e-4 mm = 100 nm — ~1/12 of a step at 800 spm, far below physical
  detection.
- De Casteljau accumulates ~6 ulp relative error in f32; at 300 mm
  that's ~215 nm absolute, well below the 1.25 µm step resolution.
- All existing host tests pass; constants relaxed to f32-achievable
  bounds in test assertions where appropriate (bench-regression test
  at 100 mm scale still passes with 1e-3 t-space tolerance).

Firmware ROM dropped from 126,732 → 117,516 bytes (~9 KB, the
soft-float lib symbols no longer linked in for the producer path).

Hardware (Trident, H7 + F446, StepTime mode): first-jog-after-restart
3 mm @ F60 now succeeds (previously crashed). Slow-medium jogs
(F60-F1200, 1-20 mm/s) all complete. High-feedrate F3000+ (50 mm/s+)
still hits the timer-in-past on bottom — separate root cause,
follow-up needed (likely command-pump throughput on F4 under
multi-segment dispatch burst, not solver CPU).

+++ rust/runtime/tests/step_time_degenerate_velocity.rs
@@ -13,8 +13,9 @@
 use runtime::step_time::{compute_next_step_time, StepTimeQuery, StepTimeResult};
 
 /// Build the four Bezier CPs of the cubic `a·u³ + b·u² + c·u + d`.
-/// Inverse of the standard Bernstein → monomial expansion.
-fn cps_from_monomial(a: f64, b: f64, c: f64, d: f64) -> [f64; 4] {
+/// Inverse of the standard Bernstein → monomial expansion. `f32` matches
+/// the runtime's MCU precision (see `bezier_root` module docs).
+fn cps_from_monomial(a: f32, b: f32, c: f32, d: f32) -> [f32; 4] {
     let p0 = d;
     let p1 = c / 3.0 + p0;
     let p2 = b / 3.0 + 2.0 * p1 - p0;
@@ -96,7 +97,7 @@ fn reverse_accel_from_rest_negative_direction() {
 #[test]
 fn truly_motionless_curve_exhausts() {
     // Constant curve x(u) = 0 everywhere.
-    let cps: [f64; 4] = [0.0, 0.0, 0.0, 0.0];
+    let cps: [f32; 4] = [0.0, 0.0, 0.0, 0.0];
     let q = StepTimeQuery {
         cps,
         step_distance: 1.0,
@@ -116,7 +117,7 @@ fn decel_to_rest_fires_all_but_last_step() {
     // step_distance = 0.5 → ~200 steps.
     // Monomial: 0·u³ + (-100)·u² + 200·u + 0.
     let cps = cps_from_monomial(0.0, -100.0, 200.0, 0.0);
-    let mut t_curr = 0.0_f64;
+    let mut t_curr = 0.0_f32;
     let mut count = 0_i32;
     loop {
         let q = StepTimeQuery {
